% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loopcityData.R
\docType{data}
\name{GM12878_10KbLoops}
\alias{GM12878_10KbLoops}
\title{GM12878 Hi-C loop calls at 10 kb resolution}
\format{
A \code{GInteractions} object with loop anchor coordinates in hg38.
}
\source{
GEO accession GSE63525 (Rao et al. 2014, Cell).
}
\usage{
data(GM12878_10KbLoops)
}
\description{
All autosomal Hi-C loop calls for GM12878 cells at 10 kb resolution,
stored as a \code{GInteractions} object.
}
\examples{
data(GM12878_10KbLoops)
GM12878_10KbLoops
}
\keyword{datasets}
