% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loopcityData.R
\docType{data}
\name{GM12878_10KbLoopsChr21}
\alias{GM12878_10KbLoopsChr21}
\title{GM12878 Hi-C loop calls on chromosome 21 (10 kb resolution)}
\format{
A \code{GInteractions} object.
}
\source{
Subset of \code{\link{GM12878_10KbLoops}}.
}
\usage{
data(GM12878_10KbLoopsChr21)
}
\description{
Chromosome 21 subset of \code{\link{GM12878_10KbLoops}}, used in package
examples and tests.
}
\examples{
data(GM12878_10KbLoopsChr21)
GM12878_10KbLoopsChr21
}
\seealso{
\code{\link{GM12878_10KbLoops}}
}
\keyword{datasets}
