% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loopcityData.R
\docType{package}
\name{loopcityData-package}
\alias{loopcityData}
\alias{loopcityData-package}
\title{loopcityData: Experimental data for the loopcity package}
\description{
Provides Hi-C loop calls and processed chromatin interaction data used in
examples and vignettes for the loopcity package.

Small datasets are shipped directly and accessible via \code{data()}.
Larger files (K562 Hi-C contact matrices, ChIP-seq bigWig tracks) are
retrieved on demand from ExperimentHub using the accessor functions below.
}
\section{Direct data objects}{

\describe{
  \item{\code{\link{GM12878_10KbLoops}}}{All GM12878 10 kb Hi-C loop calls
    as a \code{GInteractions} object.}
  \item{\code{\link{GM12878_10KbLoopsChr21}}}{Chromosome 21 subset of
    \code{GM12878_10KbLoops}.}
}
}

\section{ExperimentHub accessors}{

\describe{
  \item{\code{\link{K562_hic}}}{Path to the K562 Hi-C \code{.hic} file.}
  \item{\code{\link{K562_loops}}}{Path to the K562 5 kb loop calls
    \code{.txt} file.}
  \item{\code{\link{K562_CTCF_bw}}}{Path to K562 CTCF ChIP-seq bigWig.}
  \item{\code{\link{K562_H3K27ac_bw}}}{Path to K562 H3K27ac ChIP-seq
    bigWig.}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jpflores-13/loopcityData}
  \item Report bugs at \url{https://github.com/jpflores-13/loopcityData/issues}
}

}
\author{
\strong{Maintainer}: Sarah Parker \email{smp3800@gmail.com} (\href{https://orcid.org/0000-0002-2700-3979}{ORCID})

Authors:
\itemize{
  \item Sarah Parker \email{smp3800@gmail.com} (\href{https://orcid.org/0000-0002-2700-3979}{ORCID})
  \item JP Flores \email{jflores@unc.edu}
}

Other contributors:
\itemize{
  \item Howard Hughes Medical Institute (Gilliam Fellows Program, GT16825, to J.P.F.) [funder]
  \item National Institutes of Health (R35-GM128645, to D.H.P.) [funder]
  \item National Science Foundation (Graduate Research Fellowship Program, DGE-1650116, to S.M.P.) [funder]
}

}
